refactor(workstation): extract keyboard input handler into useInputHandler (0.72 app.ts decomposition)
